requirements.txt
django
djoser
djangorestframework
requests
mysql
mysql-connector-python
djangorestframework-simplejwt
